Feature/#109 직원 삭제 기능 수정

+++ src/main/java/kyonggi/bookslyserver/domain/shop/service/EmployeeService.java
@@ -3,6 +3,7 @@ package kyonggi.bookslyserver.domain.shop.service;
 import jakarta.transaction.Transactional;
 import kyonggi.bookslyserver.domain.shop.dto.request.EmployeeCreateRequestDto;
 import kyonggi.bookslyserver.domain.shop.dto.request.EmployeeWorkScheduleDto;
+import kyonggi.bookslyserver.domain.shop.dto.response.employee.EmployeeDeleteResponseDto;
 import kyonggi.bookslyserver.domain.shop.dto.response.employee.EmployeeCreateResponseDto;
 import kyonggi.bookslyserver.domain.shop.entity.Employee.Employee;
 import kyonggi.bookslyserver.domain.shop.entity.Employee.EmployeeMenu;
@@ -97,12 +98,17 @@ public class EmployeeService {
     }
 
     @Transactional
-    public void delete(Long id){
+    public EmployeeDeleteResponseDto delete(Long id){
         Optional<Employee> employee = employeeRepository.findById(id);
+
+        if(!employee.isPresent()){
+            throw new EntityNotFoundException();
+        }
 
         Shop shop = employee.get().getShop();
         shop.getEmployees().remove(employee.get());
         employeeRepository.deleteById(id);
+        return new EmployeeDeleteResponseDto(id);
     }
 
 }
